+++ src/main/java/com/miajon/jobsearch/model/validators/AmountNotZeroValidator.java
@@ -3,9 +3,9 @@ package com.miajon.jobsearch.model.validators;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class AmountNotZeroValidator implements ConstraintValidator<AmountNotZeroConstraint, Integer> {
+public class AmountNotZeroValidator implements ConstraintValidator<AmountNotZeroConstraint, Double> {
     @Override
-    public boolean isValid(Integer value, ConstraintValidatorContext context) {
-        return value != 0;
+    public boolean isValid(Double value, ConstraintValidatorContext context) {
+        return value != 0.;
     }
 }
